requirements.txt
Django==1.8.1
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
djangorestframework==3.1.1
kombu==3.0.26
librabbitmq==1.6.1
pytz==2015.2
wsgiref==0.1.2
